test: import data file export module normally
